front: split timezone determination to hook
