+++ app/hooks/use-location-timezone.ts
@@ -1,0 +1,30 @@
+import { EventFormSchemaType } from "@/components/form/types";
+import { currentTimezone } from "@/lib/time";
+import { useEffect, useState } from "react";
+
+export const useLocationTimezone = (location: EventFormSchemaType["location"]) => {
+  const [timezone, setTimezone] = useState<string>(currentTimezone());
+  
+  useEffect(() => {
+    const determineTimezone = async () => {
+      switch (location.kind) {
+        case "custom":
+          return location.timeZone;
+        case "virtual":
+          // ! Change to organizer timezone (#286)
+          return currentTimezone();
+        case "geo":
+          const GeoTZFind = (await import("browser-geo-tz")).find;
+          const tz = await GeoTZFind(location.lat, location.lng);
+          return tz[0];
+      }
+    }
+    
+    determineTimezone()
+      .then((found) => {
+        setTimezone(found);
+      })
+  }, [location]);
+
+  return timezone;
+};
